Move BLEU integration into the nlp folder
