+++ recipes/CVSS/S2ST/hparams/train_fr-en.yaml
@@ -220,7 +220,7 @@ test_search: !new:speechbrain.decoders.seq2seq.S2STransformerBeamSearcher
     beam_size: !ref <test_bs>
 
 acc_computer: !name:speechbrain.utils.Accuracy.AccuracyStats
-bleu_computer: !name:speechbrain.integrations.metrics.bleu.BLEUStats
+bleu_computer: !name:speechbrain.integrations.nlp.bleu.BLEUStats
     merge_words: False
 
 #checkpointer
