+++ recipes/Fisher-Callhome-Spanish/ST/transformer/hparams/transformer.yaml
@@ -245,6 +245,6 @@ checkpointer: !new:speechbrain.utils.checkpoints.Checkpointer
         normalizer: !ref <normalize>
         counter: !ref <epoch_counter>
 
-bleu_computer: !name:speechbrain.integrations.metrics.bleu.BLEUStats
+bleu_computer: !name:speechbrain.integrations.nlp.bleu.BLEUStats
     merge_words: False
 acc_computer: !name:speechbrain.utils.Accuracy.AccuracyStats
