+++ recipes/IWSLT22_lowresource/AST/transformer/hparams/train_w2v2_st.yaml
@@ -193,7 +193,7 @@ test_search: !new:speechbrain.decoders.S2STransformerBeamSearcher
     using_eos_threshold: True
     length_normalization: True
 
-bleu_computer: !name:speechbrain.integrations.metrics.bleu.BLEUStats
+bleu_computer: !name:speechbrain.integrations.nlp.bleu.BLEUStats
     merge_words: False
 
 acc_computer: !name:speechbrain.utils.Accuracy.AccuracyStats
