+++ speechbrain/utils/bleu.py
@@ -2,10 +2,10 @@
 
 import warnings
 
-from speechbrain.integrations.metrics.bleu import *  # noqa: F401, F403
+from speechbrain.integrations.nlp.bleu import *  # noqa: F401, F403
 
 warnings.warn(
-    message="speechbrain.util.bleu has moved to speechbrain.integrations.metrics.bleu",
+    message="speechbrain.util.bleu has moved to speechbrain.integrations.nlp.bleu",
     category=DeprecationWarning,
     stacklevel=2,
 )
